// server/src/graphql.schema.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: Date;
  JSON: any;
  Upload: any;
};

export type Account = {
  __typename?: 'Account';
  /** Account creation time */
  createdAt: Scalars['DateTime'];
  /** Account deletion time */
  deletedAt?: Maybe<Scalars['DateTime']>;
  /** list of documents that account is member of */
  documents: Array<AccountDocument>;
  /** Account ID */
  id: Scalars['ID'];
  /** Optional account name */
  name?: Maybe<Scalars['String']>;
  /** Account last update time */
  updatedAt: Scalars['DateTime'];
};

export type AccountDocument = {
  __typename?: 'AccountDocument';
  /** Time when account was accepted as member of document */
  acceptedAt: Scalars['DateTime'];
  /** document member creation time */
  createdAt: Scalars['DateTime'];
  /** account that created document member */
  createdBy: AccountInfo;
  /** document associated with account */
  document?: Maybe<Document>;
  /** Role of account for document */
  role: DocumentRole;
};

export type AccountInfo = {
  __typename?: 'AccountInfo';
  /** account ID */
  id: Scalars['ID'];
  /** optional account name */
  name?: Maybe<Scalars['String']>;
};

export type AccountUpdateInput = {
  /** id of the account to update */
  id: Scalars['ID'];
  /** name of the account */
  name?: InputMaybe<Scalars['String']>;
  /** update timestamp */
  timestamp: Scalars['DateTime'];
};

export type Document = {
  __typename?: 'Document';
  /** document creation time */
  createdAt: Scalars['DateTime'];
  /** account that created document */
  createdBy: AccountInfo;
  /** document deletion time */
  deletedAt?: Maybe<Scalars['DateTime']>;
  /** file associated with document */
  file?: Maybe<FileInfo>;
  /** list of highlights associated with document */
  highlights: Array<DocumentHighlight>;
  /** unique document ID */
  id: Scalars['ID'];
  /** list of document members */
  members: Array<DocumentMember>;
  /** metadata associated with document */
  meta: DocumentMeta;
  /** type of the document */
  type: DocumentType;
  /** document last update time */
  updatedAt: Scalars['DateTime'];
};
// ...
export type DocumentHighlight = {
  __typename?: 'DocumentHighlight';
  /** content associated with highlight */
  content: Scalars['JSON'];
  /** highlight creation time */
  createdAt: Scalars['DateTime'];
  /** highlight author ID */
  createdBy: AccountInfo;
  /** highlight deletion time */
  deletedAt?: Maybe<Scalars['DateTime']>;
  /** unique highlight id */
  id: Scalars['ID'];
  /** highlight location */
  location: Scalars['JSON'];
  /** highlight last udpate time */
  updatedAt: Scalars['DateTime'];
};
// ...
export type DocumentMember = {
  __typename?: 'DocumentMember';
  /** Time when account was accepted as member of document */
  acceptedAt?: Maybe<Scalars['DateTime']>;
  /** member account */
  account: AccountInfo;
  /** document member creation time */
  createdAt: Scalars['DateTime'];
  /** account that created document member */
  createdBy: AccountInfo;
  /** Role of account for document */
  role: DocumentRole;
};
// ...
export type DocumentMeta = {
  __typename?: 'DocumentMeta';
  /** document author */
  author?: Maybe<Scalars['String']>;
  /** url of document cover image */
  cover?: Maybe<Scalars['String']>;
  /** document description */
  description?: Maybe<Scalars['String']>;
  /** keywords associated with document */
  keywords?: Maybe<Array<Scalars['String']>>;
  /** number of document pages */
  pageCount?: Maybe<Scalars['Int']>;
  /** title of the document */
  title?: Maybe<Scalars['String']>;
};
// ...
export enum DocumentRole {
  Admin = 'ADMIN',
  Editor = 'EDITOR',
  Viewer = 'VIEWER'
}

export enum DocumentType {
  Pdf = 'PDF'
}

export type FileInfo = {
  __typename?: 'FileInfo';
  /** file creation time */
  createdAt: Scalars['DateTime'];
  /** account that created file */
  createdBy: AccountInfo;
  /** sha256 hash of file */
  hash: Scalars['String'];
  /** file mime type */
  mimeType: Scalars['String'];
  /** alternative file sources */
  sources?: Maybe<Array<Scalars['String']>>;
  /** file last update time */
  updatedAt: Scalars['DateTime'];
  /** primary url of file */
  url?: Maybe<Scalars['String']>;
};